+++ examples/inc/print.php
@@ -49,6 +49,9 @@ function example_print_header($title)
     example_print($title);
     example_print_separator();
     example_print();
+
+    global $timerStart;
+    $timerStart = microtime(true);
 }
 
 /**
@@ -56,7 +59,16 @@ function example_print_header($title)
  */
 function example_print_footer()
 {
+    global $timerStart;
+    $timerEnd = microtime(true);
+    $time = round(($timerEnd - $timerStart) * 1000);
+
+    $memory = memory_get_peak_usage(true) / (1024 * 1024);
+
     example_print();
+    example_print_separator();
+    example_print(sprintf(' Time   : %sms', $time));
+    example_print(sprintf(' Memory : %dMb', $memory));
     example_print_separator();
     example_print();
 }