+++ src/Client/ClientInterface.php
@@ -4,7 +4,6 @@ namespace StadGent\Services\OpeningHours\Client;
 
 use StadGent\Services\OpeningHours\Handler\HandlerInterface;
 use StadGent\Services\OpeningHours\Request\RequestInterface;
-use StadGent\Services\OpeningHours\Response\ResponseInterface;
 
 /**
  * Client Interface.